requirements.txt
alembic==0.9.2
aniso8601==1.2.1
chardet==3.0.4
click==6.7
coverage==4.4.1
feedgen==0.5.1
flake8==3.4.1
flasgger==0.8.0
Flask==0.12.2
Flask-RESTful==0.3.6
freezegun==0.3.9
gspread==0.6.2
httplib2==0.10.3
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
lxml==4.0.0
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
mypy==0.530
mysql-connector==2.1.6
pluggy==0.5.2
py==1.4.34
pycodestyle==2.3.1
pyflakes==1.5.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.18.4
rsa==3.4.2
six==1.11.0
slackweb==1.0.5
SQLAlchemy==1.1.10
sqlalchemy-seed==0.1.1
SQLAlchemy-Utils==0.32.14
tox==2.7.0
typecheck-decorator==1.3
typed-ast==1.1.0
typing==3.6.2
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
python-dotenv==0.7.1
google-api-python-client==1.6.5
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
uritemplate==3.0.0
cachetools==2.0.1
certifi==2018.1.18
google-auth==1.4.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0